Revert "fix(Form): rewrite Form.ts to .vue to fix props parsing with nuxt-component-meta"
This reverts commit 8a2176eecdaa6713a72db7d4928b94d8f4a74e9c.
